Moved BaseUri code into Requests namespace as that's what it's part of and there are 3 files now, too many for root folder
