+++ src/SevenDigital.Api.Wrapper/Extensions/ApiExtensions.cs
@@ -1,10 +1,12 @@
-﻿namespace SevenDigital.Api.Wrapper
+﻿using SevenDigital.Api.Wrapper.Requests;
+
+namespace SevenDigital.Api.Wrapper
 {
 	public static class ApiExtensions
 	{
 		public static IFluentApi<T> UsingBaseUri<T>(this IFluentApi<T> api, string baseUri)
 		{
-			api.UsingBaseUri(new StringBaseUriProvider(baseUri));
+			api.UsingBaseUri(new BaseUriFromString(baseUri));
 			return api;
 		}
 	}
