+++ src/SevenDigital.Api.Wrapper/SevenDigital.Api.Wrapper.csproj
@@ -60,8 +60,9 @@
     <Compile Include="Extensions\HasPlaylistItemIdParameterExtensions.cs" />
     <Compile Include="Extensions\ParameterExtensions.cs" />
     <Compile Include="Extensions\ShopExtensions.cs" />
-    <Compile Include="StringBaseUriProvider.cs" />
-    <Compile Include="IBaseUriProvider.cs" />
+    <Compile Include="Requests\BaseUriFromApiUri.cs" />
+    <Compile Include="Requests\BaseUriFromString.cs" />
+    <Compile Include="Requests\IBaseUriProvider.cs" />
     <Compile Include="Requests\IRequestBuilder.cs" />
     <Compile Include="Requests\RequestBuilder.cs" />
     <Compile Include="Requests\RequestPayload.cs" />
